do charset conversion on the fly * detectCharset on the html document before cleaning * perform conversion (if needed) during text extraction * decode html entities after the charset conversion * reformat entities.h (-> entities.hh) and entities.c (-> entities.cc) by changing tabs to spaces and removing carriage return
