+++ src/entities.hh
@@ -1,0 +1,25 @@
+/*	Copyright 2012 Christoph Gärtner
+	Distributed under the Boost Software License, Version 1.0
+*/
+
+#ifndef DECODE_HTML_ENTITIES_UTF8_
+#define DECODE_HTML_ENTITIES_UTF8_
+
+#include <cstddef>
+
+namespace entities {
+    std::size_t decode_html_entities_utf8(char *dest, const char *src);
+    /*	Takes input from <src> and decodes into <dest>, which should be a buffer
+        large enough to hold <strlen(src) + 1> characters.
+
+        If <src> is <NULL>, input will be taken from <dest>, decoding
+        the entities in-place.
+
+        The function returns the length of the decoded string.
+    */
+
+
+    bool simple_parse_entity(const char* from, char* to);
+}
+
+#endif
